Use the module to load from the ncm-module field
Allow to choose the Perl module to run for a given path in
/software/components.

The semantics will be as follows:

For a component defined in "/software/components/foo", if "ncm-module"
exists, that indicates the name of the component to run, so:

"/software/components/foo/ncm-module" = "bar";

will cause NCM::Component::bar to be executed.

If the subpath ncm-module doesn't exist, we'll follow the old behaviour
and run NCM::Component::foo.