ComponentProxy: disable uninitialized warnings while restoring ENV and SIG
In old perl (up to EL5), ENV and SIG are magic variables with their own warning handlers,
and you cannot assign an undef or empty string without triggering a warning.
In particular annoying to restore SIG, which is initalised as a hash with all possible
signals with undef value (and restoring those is one undef assignment per key, and thus one warning).