Allow component modules to be loaded from any path
Remove hasFile method which artificially restricts the paths which
component modules can be loaded from, allowing loading from any
valid path in `@INC`.

This allows component modules to be placed in a location appropriate
for each operating system environment we target.

This functionality may originally have been part of a mitigation
against CVE-2016-1238, but this was also addressed in 5cfe993 in a
far less restrictive way.

* Remove usage of hasFile from ComponentProxyList
* Declare mod_fn once and only when a value exists